rezerwacja szkolenia
git-svn-id: http://svn.sodexo.com.pl/svn/gryf/trunk@355 a1b3532b-92c3-dd42-b132-913ad673e08d

+++ gryf-common/src/main/java/pl/sodexo/it/gryf/common/dto/publicbenefits/traininginstiutions/detailsform/TrainingDTO.java
@@ -46,7 +46,7 @@ public class TrainingDTO extends VersionableDto {
 
     @Getter
     @Setter
-    private BigDecimal hoursNumber;
+    private Integer hoursNumber;
 
     @Getter
     @Setter
